fix: Set old system to use new sidebar, move sidebar to layout folder
