+++ frontend/src/features/layout/Sidebar/navigation/NavigationMenuList.scss
@@ -1,4 +1,4 @@
-@import '../../../../../static/css/includes/_variables_dimensions.scss';
+@import '../../../../static/css/includes/_variables_dimensions.scss';
 
 .nav-menu {
 	padding: 12px 0;
