+++ frontend/src/features/layout/Sidebar/preview/NavigationItem.stories.jsx
