+++ frontend/src/features/layout/Sidebar/preview/NavigationMenuList.stories.jsx
