+++ frontend/src/features/layout/Sidebar/preview/SidebarNavigationMenu.stories.jsx
