+++ frontend/src/features/layout/Sidebar/preview/SidebarThemeSwitcher.stories.jsx
