+++ frontend/src/features/layout/Sidebar/shell/Sidebar.scss
