+++ frontend/src/features/layout/components/AppLayout.test.jsx
@@ -11,11 +11,7 @@ vi.mock('../topbar', () => ({
 	Topbar: () => <div data-testid="topbar" />,
 }));
 
-vi.mock('../../../static/js/components/-NEW-/PageSidebar', () => ({
-	PageSidebar: () => <div data-testid="page-sidebar" />,
-}));
-
-vi.mock('../../shared/components/Sidebar', () => ({
+vi.mock('../Sidebar', () => ({
 	Sidebar: () => <div data-testid="page-sidebar" />,
 }));
