+++ frontend/src/features/layout/index.js
@@ -1,1 +1,2 @@
 export { AppLayout } from './components/AppLayout';
+export { Sidebar } from './Sidebar';
